Use double precision.
Need double precision so that exponential gain after inverse FFT does
not boost up rounding errors.

Added an option for a Ricker wavelet with specified peak frequency.

Added a demo script that compares modeled seismograms with simple
summation of time-delayed wavelets in the case where Q is infinite and
reflection coefficients are small, so that multiples are negligible.